+++ framework/contracts/account/manager/src/commands.rs
@@ -1,12 +1,12 @@
 use crate::{contract::ManagerResult, error::ManagerError, queries::query_module_cw2};
 use crate::{validation, versioning};
-use abstract_core::manager::state::ACCOUNT_FACTORY;
+use abstract_core::manager::state::{ACCOUNT_FACTORY, MODULE_QUEUE};
 
 use abstract_core::adapter::{
     AuthorizedAddressesResponse, BaseExecuteMsg, BaseQueryMsg, ExecuteMsg as AdapterExecMsg,
     QueryMsg as AdapterQuery,
 };
-use abstract_core::manager::{InternalConfigAction, QueryMsg};
+use abstract_core::manager::InternalConfigAction;
 use abstract_core::objects::gov_type::GovernanceDetails;
 use abstract_core::objects::AssetEntry;
 
@@ -36,11 +36,10 @@ use abstract_sdk::{
     ModuleRegistryInterface,
 };
 use cosmwasm_std::{
-    ensure, from_binary, to_binary, wasm_execute, Addr, Attribute, Binary, CosmosMsg, Deps,
+    ensure, from_binary, to_binary, wasm_execute, Addr, Attribute, Binary, Coin, CosmosMsg, Deps,
     DepsMut, Empty, Env, MessageInfo, Response, StdError, StdResult, Storage, SubMsg, WasmMsg,
 };
 use cw2::{get_contract_version, ContractVersion};
-use cw_ownable::Ownership;
 use cw_storage_plus::Item;
 use semver::Version;
 
@@ -83,7 +82,7 @@ pub fn update_module_addresses(
     Ok(ManagerResponse::action("update_module_addresses"))
 }
 
-// Attempts to install a new module through the Module Factory Contract
+/// Attempts to install a new module through the Module Factory Contract
 pub fn install_module(
     deps: DepsMut,
     msg_info: MessageInfo,
@@ -94,25 +93,40 @@ pub fn install_module(
     // only owner can call this method
     assert_admin_right(deps.as_ref(), &msg_info.sender)?;
 
-    // Check if module is already enabled.
-    if ACCOUNT_MODULES
-        .may_load(deps.storage, &module.id())?
-        .is_some()
-    {
-        return Err(ManagerError::ModuleAlreadyInstalled(module.id()));
-    }
-
     let config = CONFIG.load(deps.storage)?;
 
     let response =
         ManagerResponse::new("install_module", vec![("module", module.id_with_version())])
-            .add_message(wasm_execute(
+            .add_message(install_module_internal(
+                deps.storage,
+                module,
                 config.module_factory_address,
-                &ModuleFactoryMsg::InstallModule { module, init_msg },
+                init_msg,
                 msg_info.funds, // We forward all the funds to the module_factory address for them to use in the install
             )?);
 
     Ok(response)
+}
+
+/// Generate message for installing module
+pub(crate) fn install_module_internal(
+    storage: &dyn Storage,
+    module: ModuleInfo,
+    module_factory_address: Addr,
+    init_msg: Option<Binary>,
+    funds: Vec<Coin>,
+) -> ManagerResult<CosmosMsg> {
+    // Check if module is already enabled.
+    if ACCOUNT_MODULES.may_load(storage, &module.id())?.is_some() {
+        return Err(ManagerError::ModuleAlreadyInstalled(module.id()));
+    }
+
+    let msg = wasm_execute(
+        module_factory_address,
+        &ModuleFactoryMsg::InstallModule { module, init_msg },
+        funds,
+    )?;
+    Ok(msg.into())
 }
 
 // Sets the Treasury address on the module if applicable and adds it to the state
@@ -199,27 +213,30 @@ pub fn exec_on_module(
 /// Creates a sub-account for this account,
 pub fn create_sub_account(
     deps: DepsMut,
-    env: Env,
     msg_info: MessageInfo,
+    env: Env,
     name: String,
     description: Option<String>,
     link: Option<String>,
     base_asset: Option<AssetEntry>,
     namespace: Option<String>,
+    install_modules: Vec<(ModuleInfo, Option<Binary>)>,
 ) -> ManagerResult {
     // only owner can create a subaccount
     assert_admin_right(deps.as_ref(), &msg_info.sender)?;
 
     let create_account_msg = &abstract_core::account_factory::ExecuteMsg::CreateAccount {
-        /// this contract (the manager) will be the account owner
+        /// proxy of this manager will be the account owner
         governance: GovernanceDetails::SubAccount {
-            manager: env.contract.address.to_string(),
+            manager: env.contract.address.into_string(),
+            proxy: ACCOUNT_MODULES.load(deps.storage, PROXY)?.into_string(),
         },
         name,
         description,
         link,
         base_asset,
         namespace,
+        install_modules,
     };
 
     let account_factory_addr = query_module(
@@ -773,65 +790,123 @@ pub fn update_account_status(
     Ok(response)
 }
 
-pub fn update_internal_config(deps: DepsMut, info: MessageInfo, config: Binary) -> ManagerResult {
+/// Allows the owner to manually update the internal configuration of the account.
+/// This can be used to unblock the account and its modules in case of a bug/lock on the account.
+///
+/// This method is also called once by the Account Factory after the account is created and triggers any queued module installations.
+pub fn update_internal_config(
+    mut deps: DepsMut,
+    info: MessageInfo,
+    config: Binary,
+) -> ManagerResult {
+    // deserialize the config action
     let action: InternalConfigAction =
         from_binary(&config).map_err(|error| ManagerError::InvalidConfigAction { error })?;
-    match action {
-        InternalConfigAction::UpdateModuleAddresses { to_add, to_remove } => {
-            // only Account Factory/Owner can add custom modules.
-            // required to add Proxy after init by Account Factory.
-            ACCOUNT_FACTORY
-                .assert_admin(deps.as_ref(), &info.sender)
-                .or_else(|_| assert_admin_right(deps.as_ref(), &info.sender))?;
-            update_module_addresses(deps, to_add, to_remove)
+
+    let (add, remove) = match action {
+        InternalConfigAction::UpdateModuleAddresses { to_add, to_remove } => (to_add, to_remove),
+        _ => {
+            return Err(ManagerError::InvalidConfigAction {
+                error: StdError::generic_err("Unknown config action"),
+            })
         }
-        _ => Err(ManagerError::InvalidConfigAction {
-            error: StdError::generic_err("Unknown config action"),
-        }),
+    };
+
+    // if the caller is the module factory, and the account is not already instantiated, then instantiate the account and register the modules.
+    if ACCOUNT_FACTORY
+        .is_admin(deps.as_ref(), &info.sender)
+        .is_ok_and(|a| a)
+        && !ACCOUNT_MODULES.has(deps.storage, PROXY)
+    {
+        // Add the proxy.
+        update_module_addresses(deps.branch(), add, remove)?;
+        // Perform module installation.
+        let queued_modules = MODULE_QUEUE.load(deps.storage)?;
+
+        let config = CONFIG.load(deps.storage)?;
+
+        // TODO: Funds should be forwarded from account factory to this contract after init.
+        let install_msgs: ManagerResult<Vec<CosmosMsg>> = queued_modules
+            .into_iter()
+            .map(|(module, init_msg)| {
+                let module_factory_address = config.module_factory_address.clone();
+                // TODO: query fee from VC
+                let funds = vec![];
+                install_module_internal(
+                    deps.storage,
+                    module,
+                    module_factory_address,
+                    init_msg,
+                    funds,
+                )
+            })
+            .collect::<ManagerResult<Vec<CosmosMsg>>>();
+
+        // clear the queue
+        MODULE_QUEUE.remove(deps.storage);
+        // Remove account factory from storage
+        ACCOUNT_FACTORY.set(deps, None)?;
+
+        Ok(ManagerResponse::action("manager_after_init").add_messages(install_msgs?))
+    } else {
+        assert_admin_right(deps.as_ref(), &info.sender)?;
+        update_module_addresses(deps, add, remove)
     }
 }
 
-fn query_ownership(deps: Deps, maybe_manager: Addr) -> ManagerResult<String> {
-    let Ownership::<String> { owner, .. } = deps
-        .querier
-        .query_wasm_smart(maybe_manager.clone(), &QueryMsg::Ownership {})
-        .map_err(|_| ManagerError::NoContractOwner(maybe_manager.to_string()))?;
-
-    owner.ok_or(ManagerError::NoContractOwner(maybe_manager.to_string()))
-}
-
+/// Function that guards all the admin rights.
+/// This function should return `Ok` when called by:
+/// - The owner of the contract (i.e. account).
+/// - The top-level owner of the account that owns this account. I.e. the first account for which the `GovernanceDetails` is not `SubAccount`.
 fn assert_admin_right(deps: Deps, sender: &Addr) -> ManagerResult<()> {
     let ownership_test = cw_ownable::assert_owner(deps.storage, sender);
 
-    if ownership_test.is_ok() {
-        return Ok(());
+    // If the sender is the owner, the admin test is passed
+    let mut ownership_error: ManagerError = match ownership_test {
+        Ok(()) => return Ok(()),
+        Err(err) => err.into(),
+    };
+
+    // In case it fails we get the account info and check if the current(this) account is a sub-account.
+    let mut current: AccountInfo = INFO.load(deps.storage)?;
+    // Get sub-accounts until we get non-sub-account governance or reach recursion limit
+    for _ in 0..MAX_ADMIN_RECURSION {
+        match current.governance_details {
+            // As long as the accounts are sub-accounts, we check the owner of the parent account
+            GovernanceDetails::SubAccount { manager, .. } => {
+                current = INFO
+                    .query(&deps.querier, manager)
+                    .map_err(|_| ManagerError::SubAccountAdminVerification)?;
+
+                // Change error type if it was sub-account
+                ownership_error = ManagerError::SubAccountAdminVerification;
+            }
+            _ => break,
+        }
     }
 
-    let account_info = INFO.load(deps.storage)?;
-    match account_info.governance_details {
-        // If the account has SubAccount governance, the owner of the manager also has admin rights over this account
-        GovernanceDetails::SubAccount { manager } => {
-            // We try to query the ownership of the manager monarch account if the first query failed
-            let mut current = manager;
-            let mut i = 0;
-            while i < MAX_ADMIN_RECURSION {
-                let owner = query_ownership(deps, current)
-                    .map_err(|_| ManagerError::SubAccountAdminVerification)?;
-                if *sender == owner {
-                    // If the owner of the current contract is the sender, the admin test is passed
-                    return Ok(());
-                } else {
-                    // If not, we try again with the queried owner
-                    current = deps.api.addr_validate(&owner)?
-                }
-                i += 1;
+    match current.governance_details {
+        GovernanceDetails::Monarchy { monarch: owner }
+        | GovernanceDetails::External {
+            governance_address: owner,
+            ..
+        } => {
+            // If the owner of the top-level account is the sender, the admin test is passed.
+            // This gives the top-level owner the ability to manage all sub-accounts.
+            if *sender == owner {
+                Ok(())
+            } else {
+                Err(ownership_error)
             }
+        }
+        // MAX_ADMIN_RECURSION levels deep still sub account
+        GovernanceDetails::SubAccount { .. } => {
             Err(ManagerError::Std(StdError::generic_err(format!(
-                "Admin recursion error, too much recursion, maximum allowed admin recursion : {}",
+                "Admin recursion error, too much recursion, maximum allowed sub-account admin recursion : {}",
                 MAX_ADMIN_RECURSION
             ))))
         }
-        _ => Ok(ownership_test?),
+        _ => Err(ownership_error),
     }
 }
 
@@ -888,7 +963,7 @@ mod tests {
         let _info = mock_info("not_owner", &[]);
 
         let res = execute_as(deps.as_mut(), "not_owner", msg);
-        assert_that(&res)
+        assert_that!(&res)
             .is_err()
             .is_equal_to(ManagerError::Ownership(OwnershipError::NotOwner));
 
@@ -949,14 +1024,14 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), set_owner_msg);
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
             let accept_msg = ExecuteMsg::UpdateOwnership(cw_ownable::Action::AcceptOwnership);
             execute_as(deps.as_mut(), new_owner, accept_msg)?;
 
             let actual_owner = cw_ownable::get_ownership(&deps.storage)?.owner.unwrap();
 
-            assert_that(&actual_owner).is_equal_to(Addr::unchecked(new_owner));
+            assert_that!(&actual_owner).is_equal_to(Addr::unchecked(new_owner));
 
             Ok(())
         }
@@ -975,7 +1050,7 @@ mod tests {
             execute_as_owner(deps.as_mut(), msg)?;
 
             let actual_info = INFO.load(deps.as_ref().storage)?;
-            assert_that(&actual_info.governance_details.owner_address().to_string())
+            assert_that!(&actual_info.governance_details.owner_address().to_string())
                 .is_equal_to("new_gov".to_string());
 
             Ok(())
@@ -997,17 +1072,17 @@ mod tests {
             ];
 
             let res = update_module_addresses(deps.as_mut(), Some(to_add.clone()), Some(vec![]));
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
             let actual_modules = load_account_modules(&deps.storage)?;
 
             speculoos::prelude::VecAssertions::has_length(
-                &mut assert_that(&actual_modules),
+                &mut assert_that!(&actual_modules),
                 to_add.len(),
             );
             for (module_id, addr) in to_add {
                 speculoos::iter::ContainingIntoIterAssertions::contains(
-                    &mut assert_that(&actual_modules),
+                    &mut assert_that!(&actual_modules),
                     &(module_id, Addr::unchecked(addr)),
                 );
             }
@@ -1023,7 +1098,7 @@ mod tests {
             let to_add: Vec<(String, String)> = vec![("".to_string(), "module1_addr".to_string())];
 
             let res = update_module_addresses(deps.as_mut(), Some(to_add), Some(vec![]));
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::InvalidModuleName {});
 
@@ -1045,11 +1120,11 @@ mod tests {
             let to_remove: Vec<String> = vec!["test:module".to_string()];
 
             let res = update_module_addresses(deps.as_mut(), Some(vec![]), Some(to_remove));
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
             let actual_modules = load_account_modules(&deps.storage)?;
 
-            speculoos::prelude::VecAssertions::is_empty(&mut assert_that(&actual_modules));
+            speculoos::prelude::VecAssertions::is_empty(&mut assert_that!(&actual_modules));
 
             Ok(())
         }
@@ -1062,7 +1137,7 @@ mod tests {
             let to_remove: Vec<String> = vec![PROXY.to_string()];
 
             let res = update_module_addresses(deps.as_mut(), Some(vec![]), Some(to_remove));
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::CannotRemoveProxy {});
 
@@ -1073,21 +1148,36 @@ mod tests {
         fn only_account_factory_or_owner() -> ManagerTestResult {
             let mut deps = mock_dependencies();
             mock_init(deps.as_mut())?;
-
-            let action = InternalConfigAction::UpdateModuleAddresses {
-                to_add: None,
+            // mock add the proxy
+            let action_add = InternalConfigAction::UpdateModuleAddresses {
+                to_add: Some(vec![(PROXY.to_string(), "module_addr".to_string())]),
                 to_remove: None,
             };
-            let msg = ExecuteMsg::UpdateInternalConfig(to_binary(&action).unwrap());
+            let msg = ExecuteMsg::UpdateInternalConfig(to_binary(&action_add).unwrap());
 
             let res = execute_as(deps.as_mut(), TEST_ACCOUNT_FACTORY, msg.clone());
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
+            // add some other thing
+            let action_add = InternalConfigAction::UpdateModuleAddresses {
+                to_add: Some(vec![(
+                    "module:other".to_string(),
+                    "module_addr".to_string(),
+                )]),
+                to_remove: None,
+            };
+            let msg = ExecuteMsg::UpdateInternalConfig(to_binary(&action_add).unwrap());
+
+            // the factory can not call this
+            let res = execute_as(deps.as_mut(), TEST_ACCOUNT_FACTORY, msg.clone());
+            assert_that!(&res).is_err();
+
+            // only the owner can
             let res = execute_as_owner(deps.as_mut(), msg.clone());
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
             let res = execute_as(deps.as_mut(), "not_account_factory", msg);
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::Ownership(OwnershipError::NotOwner {}));
 
@@ -1109,7 +1199,7 @@ mod tests {
             };
 
             let res = execute_as(deps.as_mut(), "not_owner", msg);
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::Ownership(OwnershipError::NotOwner));
 
@@ -1134,7 +1224,7 @@ mod tests {
             )?;
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_err().matches(|e| {
+            assert_that!(&res).is_err().matches(|e| {
                 let _module_id = String::from("test:module");
                 matches!(e, ManagerError::ModuleAlreadyInstalled(_module_id))
             });
@@ -1153,7 +1243,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
             Ok(())
         }
@@ -1172,7 +1262,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
             let msgs = res.unwrap().messages;
 
@@ -1187,9 +1277,9 @@ mod tests {
                 vec![],
             )?
             .into();
-            assert_that(&msgs).has_length(1);
+            assert_that!(&msgs).has_length(1);
 
-            assert_that(&msg.msg).is_equal_to(&expected_msg);
+            assert_that!(&msg.msg).is_equal_to(&expected_msg);
 
             Ok(())
         }
@@ -1224,7 +1314,7 @@ mod tests {
             DEPENDENTS.save(&mut deps.storage, test_module, &dependents)?;
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::ModuleHasDependents(Vec::from_iter(
                     dependents,
@@ -1243,7 +1333,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::CannotRemoveProxy {});
 
@@ -1277,7 +1367,7 @@ mod tests {
             };
 
             let res = execute_as(deps.as_mut(), "not_module_factory", msg);
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::CallerNotModuleFactory {});
 
@@ -1310,7 +1400,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::ModuleNotFound(missing_module));
 
@@ -1377,13 +1467,13 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
             let info = INFO.load(deps.as_ref().storage)?;
 
-            assert_that(&info.name).is_equal_to(name.to_string());
-            assert_that(&info.description.unwrap()).is_equal_to(description.to_string());
-            assert_that(&info.link.unwrap()).is_equal_to(link.to_string());
+            assert_that!(&info.name).is_equal_to(name.to_string());
+            assert_that!(&info.description.unwrap()).is_equal_to(description.to_string());
+            assert_that!(&info.link.unwrap()).is_equal_to(link.to_string());
 
             Ok(())
         }
@@ -1414,13 +1504,13 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
             let info = INFO.load(deps.as_ref().storage)?;
 
-            assert_that(&info.name).is_equal_to(&prev_name);
-            assert_that(&info.description).is_none();
-            assert_that(&info.link).is_none();
+            assert_that!(&info.name).is_equal_to(&prev_name);
+            assert_that!(&info.description).is_none();
+            assert_that!(&info.link).is_none();
 
             Ok(())
         }
@@ -1437,7 +1527,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_err().matches(|e| {
+            assert_that!(&res).is_err().matches(|e| {
                 matches!(
                     e,
                     ManagerError::Validation(ValidationError::TitleInvalidShort(_))
@@ -1451,7 +1541,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_err().matches(|e| {
+            assert_that!(&res).is_err().matches(|e| {
                 matches!(
                     e,
                     ManagerError::Validation(ValidationError::TitleInvalidLong(_))
@@ -1473,7 +1563,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_err().matches(|e| {
+            assert_that!(&res).is_err().matches(|e| {
                 matches!(
                     e,
                     ManagerError::Validation(ValidationError::LinkInvalidShort(_))
@@ -1487,7 +1577,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_err().matches(|e| {
+            assert_that!(&res).is_err().matches(|e| {
                 matches!(
                     e,
                     ManagerError::Validation(ValidationError::LinkInvalidLong(_))
@@ -1532,7 +1622,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::ModuleNotFound(IBC_CLIENT.to_string()));
 
@@ -1551,7 +1641,7 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .matches(|e| matches!(e, ManagerError::ModuleAlreadyInstalled(_)));
 
@@ -1570,10 +1660,10 @@ mod tests {
             };
 
             let res = execute_as_owner(deps.as_mut(), msg);
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
 
             let msgs = res.unwrap().messages;
-            assert_that(&msgs).has_length(1);
+            assert_that!(&msgs).has_length(1);
 
             let msg = &msgs[0];
 
@@ -1585,7 +1675,7 @@ mod tests {
                 vec![],
             )?
             .into();
-            assert_that(&msg.msg).is_equal_to(&expected_msg);
+            assert_that!(&msg.msg).is_equal_to(&expected_msg);
 
             Ok(())
         }
@@ -1613,7 +1703,7 @@ mod tests {
                 msg,
             );
 
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .matches(|err| matches!(err, ManagerError::Std(StdError::GenericErr { .. })));
 
@@ -1648,7 +1738,7 @@ mod tests {
             let res = execute_as_owner(deps.as_mut(), msg);
             assert_that!(res).is_ok();
             let actual_is_suspended = SUSPENSION_STATUS.load(&deps.storage).unwrap();
-            assert_that(&actual_is_suspended).is_true();
+            assert_that!(&actual_is_suspended).is_true();
 
             let update_info_msg = ExecuteMsg::UpdateInfo {
                 name: Some("asonetuh".to_string()),
@@ -1658,7 +1748,7 @@ mod tests {
 
             let res = execute_as_owner(deps.as_mut(), update_info_msg);
 
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::AccountSuspended {});
 
@@ -1676,9 +1766,9 @@ mod tests {
 
             let res = execute_as_owner(deps.as_mut(), msg);
 
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
             let actual_is_suspended = SUSPENSION_STATUS.load(&deps.storage).unwrap();
-            assert_that(&actual_is_suspended).is_true();
+            assert_that!(&actual_is_suspended).is_true();
             Ok(())
         }
 
@@ -1693,9 +1783,9 @@ mod tests {
 
             let res = execute_as_owner(deps.as_mut(), msg);
 
-            assert_that(&res).is_ok();
+            assert_that!(&res).is_ok();
             let actual_status = SUSPENSION_STATUS.load(&deps.storage).unwrap();
-            assert_that(&actual_status).is_false();
+            assert_that!(&actual_status).is_false();
             Ok(())
         }
     }
@@ -1721,15 +1811,15 @@ mod tests {
             let bad_sender = "not_account_factory";
             let res = execute_as(deps.as_mut(), bad_sender, msg.clone());
 
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .is_equal_to(ManagerError::Ownership(OwnershipError::NotOwner {}));
 
             let owner_res = execute_as_owner(deps.as_mut(), msg.clone());
-            assert_that(&owner_res).is_ok();
+            assert_that!(&owner_res).is_ok();
 
             let factory_res = execute_as(deps.as_mut(), TEST_ACCOUNT_FACTORY, msg);
-            assert_that(&factory_res).is_ok();
+            assert_that!(&factory_res).is_ok();
 
             Ok(())
         }
@@ -1743,7 +1833,7 @@ mod tests {
 
             let res = execute_as(deps.as_mut(), TEST_ACCOUNT_FACTORY, msg);
 
-            assert_that(&res)
+            assert_that!(&res)
                 .is_err()
                 .matches(|e| matches!(e, ManagerError::InvalidConfigAction { .. }));